[Refactor] Extract docker related functions
