+++ src/main/kotlin/io/ghostbuster91/docker/compose/logger/LoggerApp.kt
@@ -4,8 +4,9 @@ import com.googlecode.lanterna.input.KeyStroke
 import com.googlecode.lanterna.input.KeyType
 import com.googlecode.lanterna.terminal.DefaultTerminalFactory
 import com.googlecode.lanterna.terminal.Terminal
-import de.gesellix.docker.compose.ComposeFileReader
 import io.ghostbuster91.docker.compose.logger.Colors.red
+import io.ghostbuster91.docker.compose.logger.docker.readServicesFromDockerConfig
+import io.ghostbuster91.docker.compose.logger.docker.streamFromDockerCompose
 import io.ghostbuster91.docker.compose.logger.keyboard.EffectsMapping
 import io.ghostbuster91.docker.compose.logger.keyboard.KeyboardLayout
 import io.ghostbuster91.docker.compose.logger.keyboard.ServiceMappingImpl
@@ -15,10 +16,6 @@ import io.reactivex.Emitter
 import io.reactivex.Flowable
 import io.reactivex.functions.Consumer
 import io.reactivex.schedulers.Schedulers
-import java.io.BufferedReader
-import java.io.File
-import java.io.FileInputStream
-import java.io.UncheckedIOException
 
 typealias Service = String
 
@@ -40,34 +37,8 @@ fun main(args: Array<String>) {
                     }
                 }
 
-        val servicesStream = keyStrokesStream
-                .doOnSubscribe { consoleOutConsumer.accept("Press any key to continue...") }
-                .scan(StreamDefinition(setOf(terminal.waitForNumberInput(keyboardLayout).service))) { acc: StreamDefinition, keyStroke: KeyStroke ->
-                    val character = keyStroke.character?.toString()
-                    if (acc.find != null) {
-                        when {
-                            keyStroke.keyType == KeyType.Escape -> acc.copy(find = null)
-                            keyStroke.keyType == KeyType.Backspace -> acc.copy(find = acc.find.copy(text = acc.find.text.dropLast(1)))
-                            character != null && keyStroke.keyType != KeyType.Enter-> acc.copy(find = acc.find.copy(text = acc.find.text + character))
-                            else -> acc
-                        }
-                    } else {
-                        val userCommand = keyboardLayout.mapCommand(keyStroke)
-                        when (userCommand) {
-                            is UserCommand.Add -> StreamDefinition(acc.services + userCommand.service)
-                            is UserCommand.Single -> StreamDefinition(setOf(userCommand.service))
-                            is UserCommand.Remove -> StreamDefinition(acc.services - userCommand.service)
-                            UserCommand.Control.SwitchTimestamp -> acc.copy(showTimeStamps = !acc.showTimeStamps)
-                            UserCommand.Control.SwitchHelp -> acc.copy(showHelp = !acc.showHelp)
-                            UserCommand.Control.SwitchFind -> acc.copy(find = FindOption(""))
-                            else -> acc
-                        }
-                    }
-                }
-                .distinctUntilChanged()
-                .doOnNext { consoleOutConsumer.accept("Switched to $it") }
-
-        servicesStream
+        keyStrokesStream
+                .serviceStream(consoleOutConsumer, terminal, keyboardLayout)
                 .switchMap { serviceDef ->
                     val stream = if (serviceDef.showHelp) {
                         emitHelp(keyboardLayout)
@@ -76,12 +47,48 @@ fun main(args: Array<String>) {
                     } else {
                         streamFromDockerCompose(serviceDef)
                     }
-                    serviceDef.find
+                    serviceDef.search
                             ?.filter(stream)
-                            ?.map { if(serviceDef.find.text.isNotEmpty()) it.replace(serviceDef.find.text, serviceDef.find.text.red()) else it }
+                            ?.map { if (serviceDef.search.text.isNotEmpty()) it.replace(serviceDef.search.text, serviceDef.search.text.red()) else it }
                             ?: stream
                 }
                 .blockingSubscribe(consoleOutConsumer)
+    }
+}
+
+private fun Flowable<KeyStroke>.serviceStream(consoleOutConsumer: ConsoleOutConsumer, terminal: Terminal, keyboardLayout: LinuxKeyboardLayout): Flowable<StreamDefinition> {
+    return doOnSubscribe { consoleOutConsumer.accept("Press any key to continue...") }
+            .scan(StreamDefinition(setOf(terminal.waitForNumberInput(keyboardLayout).service))) { acc: StreamDefinition, keyStroke: KeyStroke ->
+                val character = keyStroke.character?.toString()
+                if (acc.search != null) {
+                    handleSearchInput(keyStroke, acc, character)
+                } else {
+                    handleUserCommand(keyboardLayout, keyStroke, acc)
+                }
+            }
+            .distinctUntilChanged()
+            .doOnNext { consoleOutConsumer.accept("Switched to $it") }
+}
+
+private fun handleSearchInput(keyStroke: KeyStroke, acc: StreamDefinition, character: String?): StreamDefinition {
+    return when {
+        keyStroke.keyType == KeyType.Escape -> acc.copy(search = null)
+        keyStroke.keyType == KeyType.Backspace -> acc.copy(search = acc.search!!.copy(text = acc.search.text.dropLast(1)))
+        character != null && keyStroke.keyType != KeyType.Enter -> acc.copy(search = acc.search!!.copy(text = acc.search.text + character))
+        else -> acc
+    }
+}
+
+private fun handleUserCommand(keyboardLayout: LinuxKeyboardLayout, keyStroke: KeyStroke, acc: StreamDefinition): StreamDefinition {
+    val userCommand = keyboardLayout.mapCommand(keyStroke)
+    return when (userCommand) {
+        is UserCommand.Add -> StreamDefinition(acc.services + userCommand.service)
+        is UserCommand.Single -> StreamDefinition(setOf(userCommand.service))
+        is UserCommand.Remove -> StreamDefinition(acc.services - userCommand.service)
+        UserCommand.Control.SwitchTimestamp -> acc.copy(showTimeStamps = !acc.showTimeStamps)
+        UserCommand.Control.SwitchHelp -> acc.copy(showHelp = !acc.showHelp)
+        UserCommand.Control.SwitchSearchFor -> acc.copy(search = SearchFor(""))
+        else -> acc
     }
 }
 
@@ -115,13 +122,6 @@ private fun Terminal.waitForNumberInput(keyboardLayout: KeyboardLayout): UserCom
             .blockingFirst()
 }
 
-private fun readServicesFromDockerConfig(args: Array<String>): List<String> {
-    val workDir = System.getProperty("user.dir")
-    val inputStream = FileInputStream(File(workDir, args.getOrElse(0) { "docker-compose.yml" }))
-    val config = ComposeFileReader().loadYaml(inputStream)
-    return config["services"]!!.keys.toList()
-}
-
 private fun userSingleStream(userInput: Flowable<KeyStroke>, singleServiceMapping: SingleServiceMapping): Flowable<UserCommand.Single> {
     return userInput
             .filter { singleServiceMapping.isSingleServiceKey(it) }
@@ -140,42 +140,6 @@ private fun keyStrokeStream(terminal: Terminal): Flowable<KeyStroke> {
     }
 }
 
-private fun streamFromDockerCompose(streamDefinition: StreamDefinition): Flowable<String> {
-    val command = DockerComposeCommandBuilder(
-            services = streamDefinition.services.toList(),
-            showTimestamps = streamDefinition.showTimeStamps)
-            .build()
-    return createProcess(command)
-            .mergeWith(Flowable.never())
-            .subscribeOn(Schedulers.io())
-}
-
-private fun createProcess(command: Command): Flowable<String> {
-    return Flowable
-            .using(
-                    {
-                        ProcessBuilder()
-                                .directory(File(System.getProperty("user.dir")))
-                                .redirectErrorStream(true)
-                                .command(command)
-                                .start()
-                    },
-                    { process -> process.inputStream.bufferedReader().toFlowable() },
-                    { process -> process.destroy() }
-            )
-            .onErrorResumeNext { t: Throwable -> if (t is UncheckedIOException) Flowable.empty<String>() else throw t }
-
-}
-
-fun BufferedReader.toFlowable(): Flowable<String> {
-    return Flowable.fromIterable(object : Iterable<String> {
-        override fun iterator(): Iterator<String> {
-            return lines().iterator()
-        }
-    })
-}
-
-typealias Command = List<String>
 
 sealed class Action {
     object Unknown : Action()
@@ -188,7 +152,7 @@ sealed class UserCommand : Action() {
     sealed class Control : UserCommand() {
         object SwitchTimestamp : Control()
         object SwitchHelp : Control()
-        object SwitchFind : Control()
+        object SwitchSearchFor : Control()
     }
 }
 
@@ -204,28 +168,8 @@ sealed class Effect : Action() {
 data class StreamDefinition(val services: Set<String>,
                             val showTimeStamps: Boolean = false,
                             val showHelp: Boolean = false,
-                            val find: FindOption? = null)
+                            val search: SearchFor? = null)
 
-data class FindOption(val text: String)
+data class SearchFor(val text: String)
 
-fun FindOption.filter(input: Flowable<String>) = input.filter { it.contains(text) }
-
-data class DockerComposeCommandBuilder(val services: List<String>,
-                                       val showTimestamps: Boolean = false,
-                                       val showColors: Boolean = true,
-                                       val follow: Boolean = true,
-                                       val tailSize: Int = 200) {
-    fun build(): Command {
-        var command = listOf("docker-compose", "logs", "--tail=$tailSize")
-        if (showTimestamps) {
-            command += "--timestamps"
-        }
-        if (!showColors) {
-            command += "--no-color"
-        }
-        if (follow) {
-            command += "--follow"
-        }
-        return command + services
-    }
-}
+private fun SearchFor.filter(input: Flowable<String>) = input.filter { it.contains(text) }
